fix: support for new activity types
There isn't any API documentation for the Xbox APIs I'm using, so
I learn about new activity types when I see one in the wild.

Saw a "LegacyAchievement" activity event yesterday, so this adds
support for it. It looks similar to an achievement, other than it
doesn't have an icon, so I'm just reusing the schema for that.

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>